+++ src/dune_rules/expander.mli
@@ -19,6 +19,11 @@ val dir : t -> Path.Build.t
 
 val context : t -> Context.t
 
+(** local or installed package *)
+type any_package =
+  | Local of Package.t
+  | Installed of Dune_package.t
+
 val make :
      scope:Scope.t
   -> scope_host:Scope.t
@@ -26,7 +31,7 @@ val make :
   -> lib_artifacts:Artifacts.Public_libs.t
   -> lib_artifacts_host:Artifacts.Public_libs.t
   -> bin_artifacts_host:Artifacts.Bin.t
-  -> find_package:(Package.Name.t -> Package.t option)
+  -> find_package:(Package.Name.t -> any_package option)
   -> t
 
 val set_foreign_flags :
@@ -124,4 +129,4 @@ val map_exe : t -> Path.t -> Path.t
 
 val artifacts : t -> Artifacts.Bin.t
 
-val find_package : t -> Package.Name.t -> Package.t option
+val find_package : t -> Package.Name.t -> any_package option