+++ test/blackbox-tests/test-cases/github1549.t/run.t
@@ -6,6 +6,24 @@ Reproduction case for #1549: too many parentheses in installed .dune files
   $ dune_cmd cat backend/_build/install/default/lib/dune_inline_tests/dune-package | sed "s/(lang dune .*)/(lang dune <version>)/" | dune_cmd sanitize
   (lang dune <version>)
   (name dune_inline_tests)
+  (sections
+   (lib
+    $TESTCASE_ROOT/backend/_build/install/default/lib/dune_inline_tests)
+   (libexec
+    $TESTCASE_ROOT/backend/_build/install/default/lib/dune_inline_tests))
+  (files
+   (lib
+    (META
+     dune-package
+     opam
+     simple_tests.ml
+     simple_tests.cmi
+     simple_tests.cmx
+     simple_tests.cmt
+     simple_tests.cma
+     simple_tests.cmxa
+     simple_tests$ext_lib))
+   (libexec (simple_tests.cmxs)))
   (library
    (name dune_inline_tests)
    (kind normal)